Move ElementDump into Dom subnamespace
